Implementing Attibutes modal as component
